test(sandbox): dedupe fake test models/tool base and document stateless runner behavior
